+++ util.h
@@ -42,7 +42,7 @@ void get_current_time(char* str, const int n, const char* format);
 /*
 * Replaces all placeholders inside raw_message and returns a pointer to the updated string (which must be free'd).
 */
-char* insert_placeholders(const char* raw_message, const connectivity_check_t* check, const enum run_if state, const struct timespec previous_last_reply, const char* datetime_format, const double diff);
+char* insert_placeholders(const char* raw_message, const connectivity_check_t* check, const enum run_if state, const struct timespec previous_last_reply, const char* datetime_format, const double diff, const int connected);
 
 /*
 * Calculates the difference in seconds of old and new