+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * 0.0.5
     * new ERROR loglevel (nothing logged regarding connections)
     * Fix free when using a reboot action
+    * Added `datetime_format` as a configuration in srd.conf
 
 * 0.0.4
     * Fix some memory not getting deallocated